pyproject.toml
[tool.poetry]
name = "stock-bag-api"
version = "0.1.0"
description = ""
authors = ["Nurlan-Aliev <nualiev612@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.32"
pydantic-settings = "^2.4.0"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
alembic = "^1.13.2"
greenlet = "^3.0.3"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
